# backend/app/api/v1/repos.py
"""
Repository Endpoints.

CRUD operations for monitored repositories.
"""

from typing import Any

from fastapi import APIRouter, HTTPException
from pydantic import BaseModel, HttpUrl

router = APIRouter()


class RepoCreate(BaseModel):
    """Schema for creating a new repository."""

    url: HttpUrl
    name: str | None = None
    description: str | None = None


class RepoResponse(BaseModel):
    """Schema for repository response."""

    id: str
    url: str
    name: str
    description: str | None
    status: str


@router.get("")
async def list_repos() -> dict[str, Any]:
    """
    List all monitored repositories.

    Returns:
        List of repositories.
    """
    # TODO: Implement database query
    return {
        "data": [],
        "total": 0,
        "page": 1,
        "page_size": 20,
    }


@router.post("", response_model=RepoResponse, status_code=201)
async def create_repo(repo: RepoCreate) -> RepoResponse:
    """
    Add a new repository to monitor.

    Args:
        repo: Repository creation data.

    Returns:
        Created repository.
    """
    # TODO: Implement database insert
    return RepoResponse(
        id="placeholder-id",
        url=str(repo.url),
        name=repo.name or "Unnamed",
        description=repo.description,
        status="pending",
    )


@router.get("/{repo_id}", response_model=RepoResponse)
async def get_repo(repo_id: str) -> RepoResponse:
    """
    Get a specific repository by ID.

    Args:
        repo_id: Repository ID.

    Returns:
        Repository details.
    """
    # TODO: Implement database query
    raise HTTPException(status_code=404, detail="Repository not found")


@router.delete("/{repo_id}", status_code=204)
async def delete_repo(repo_id: str) -> None:
    """
    Remove a repository from monitoring.

    Args:
        repo_id: Repository ID.
    """
    # TODO: Implement database delete
    raise HTTPException(status_code=404, detail="Repository not found")