# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: DevBridge API
  version: 1.0.0
  description: |
    API for DevBridge - Making Technical Work Visible to Non-Technical Stakeholders.
    
    ## Authentication
    All endpoints except `/health` require a Bearer token obtained via GitHub OAuth.
    
    ## Rate Limiting
    - 100 requests per minute per API key
    - 1000 requests per hour per API key
  contact:
    name: DevBridge Team
    email: api@devbridge.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8000/api/v1
    description: Development server
  - url: https://api.devbridge.io/v1
    description: Production server

tags:
  - name: Health
    description: Health check endpoints
  - name: Repositories
    description: Repository management
  - name: Commits
    description: Commit analysis
  - name: Translations
    description: Business translations
  - name: Chat
    description: RAG chat interface

paths:
  /health:
    get:
      tags: [Health]
      summary: Health check
      operationId: getHealth
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /ready:
    get:
      tags: [Health]
      summary: Readiness check
      operationId: getReady
      responses:
        "200":
          description: All services ready
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
        "503":
          description: Service unavailable
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"

  /repos:
    get:
      tags: [Repositories]
      summary: List repositories
      operationId: listRepositories
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
      responses:
        "200":
          description: List of repositories
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RepositoryListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Repositories]
      summary: Add a repository
      operationId: createRepository
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RepositoryCreate"
      responses:
        "201":
          description: Repository created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"

  /repos/{repoId}:
    get:
      tags: [Repositories]
      summary: Get repository details
      operationId: getRepository
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RepoIdParam"
      responses:
        "200":
          description: Repository details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Repository"
        "404":
          $ref: "#/components/responses/NotFound"

  /repos/{repoId}/commits:
    get:
      tags: [Commits]
      summary: List commits for a repository
      operationId: listCommits
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/RepoIdParam"
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: since
          in: query
          schema:
            type: string
            format: date-time
          description: Filter commits after this date
        - name: author
          in: query
          schema:
            type: string
          description: Filter by author
      responses:
        "200":
          description: List of commits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommitListResponse"

  /translations:
    get:
      tags: [Translations]
      summary: List business translations
      operationId: listTranslations
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PerPageParam"
        - name: minConfidence
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100
          description: Minimum confidence score
      responses:
        "200":
          description: List of translations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TranslationListResponse"

  /chat:
    post:
      tags: [Chat]
      summary: Send chat message
      operationId: sendChatMessage
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatRequest"
      responses:
        "200":
          description: Chat response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatResponse"
        "400":
          $ref: "#/components/responses/BadRequest"

  /webhooks/github:
    post:
      tags: [Webhooks]
      summary: GitHub webhook endpoint
      operationId: handleGitHubWebhook
      parameters:
        - name: X-GitHub-Event
          in: header
          required: true
          schema:
            type: string
        - name: X-Hub-Signature-256
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Webhook processed
        "400":
          description: Invalid signature

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  parameters:
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: perPage
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    RepoIdParam:
      name: repoId
      in: path
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ApiError"

  schemas:
    HealthStatus:
      type: object
      required: [status, version, timestamp]
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        version:
          type: string
        timestamp:
          type: string
          format: date-time
        services:
          type: array
          items:
            $ref: "#/components/schemas/ServiceHealth"

    ServiceHealth:
      type: object
      required: [name, status]
      properties:
        name:
          type: string
        status:
          type: string
          enum: [up, down, degraded]
        latencyMs:
          type: integer
        message:
          type: string

    Repository:
      type: object
      required: [id, url, name, owner, createdAt]
      properties:
        id:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        name:
          type: string
        owner:
          type: string
        description:
          type: string
        defaultBranch:
          type: string
          default: main
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    RepositoryCreate:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: "https://github.com/owner/repo"
        description:
          type: string

    RepositoryListResponse:
      type: object
      required: [items, total, page, perPage]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Repository"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer

    Commit:
      type: object
      required: [sha, message, author, timestamp]
      properties:
        sha:
          type: string
        message:
          type: string
        author:
          type: string
        authorEmail:
          type: string
          format: email
        timestamp:
          type: string
          format: date-time
        filesChanged:
          type: integer
        additions:
          type: integer
        deletions:
          type: integer

    CommitListResponse:
      type: object
      required: [items, total, page, perPage]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Commit"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer

    BusinessTranslation:
      type: object
      required: [id, commitSha, title, technicalSummary, businessValue, confidenceScore]
      properties:
        id:
          type: string
          format: uuid
        commitSha:
          type: string
        title:
          type: string
        technicalSummary:
          type: string
        businessValue:
          type: string
        risksMitigated:
          type: array
          items:
            type: string
        alignedPillars:
          type: array
          items:
            type: string
        confidenceScore:
          type: integer
          minimum: 0
          maximum: 100
        createdAt:
          type: string
          format: date-time

    TranslationListResponse:
      type: object
      required: [items, total, page, perPage]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/BusinessTranslation"
        total:
          type: integer
        page:
          type: integer
        perPage:
          type: integer

    ChatRequest:
      type: object
      required: [message]
      properties:
        message:
          type: string
          example: "O que o time fez essa semana?"
        repositoryId:
          type: string
          format: uuid
        conversationId:
          type: string

    ChatResponse:
      type: object
      required: [message, conversationId]
      properties:
        message:
          $ref: "#/components/schemas/ChatMessage"
        conversationId:
          type: string

    ChatMessage:
      type: object
      required: [id, role, content, timestamp]
      properties:
        id:
          type: string
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        timestamp:
          type: string
          format: date-time

    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          example: "DVB-001"
        message:
          type: string
          example: "Invalid request parameters"
        details:
          type: object